common page 작업 - 1  - api 2개 추가 (정리)   -- /api/posting/top   -- /api/category/top
