+++ src/main/java/com/jhpark/marketing/blog/controller/rest/category/CategoryRestController.java
@@ -1,4 +1,4 @@
-package com.jhpark.marketing.blog.controller.view.category;
+package com.jhpark.marketing.blog.controller.rest.category;
 
 import com.jhpark.marketing.blog.entity.Category;
 import com.jhpark.marketing.blog.entity.Posting;
@@ -7,6 +7,7 @@ import com.jhpark.marketing.blog.service.category.CategoryService;
 import com.jhpark.marketing.blog.service.posting.PostingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Slice;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -20,13 +21,13 @@ import java.util.Optional;
 @RequestMapping(path = "/api/category", produces = "application/json")
 @RequiredArgsConstructor
 public class CategoryRestController {
-  private final PostingService postingService;
+
   private final CategoryService categoryService;
 
-  @RequestMapping(path = "/list", method = RequestMethod.GET)
-  public List<CategoryListElementResponse> categoryList() {
-    List<CategoryListElementResponse> categories = postingService.getCategoryGroupByCount();
+  @RequestMapping(path = "/top", method = RequestMethod.GET)
+  public Slice<CategoryListElementResponse> topCategoryListByView(@RequestParam(value = "count", required = false, defaultValue = "5") int count) {
+    Slice<CategoryListElementResponse> categories = categoryService.getTopCategoryByView(count);
     return categories;
-  }
+   }
 
 }
