+++ src/main/resources/templates/posting/post_gallery_rights.ftl
@@ -1,6 +1,6 @@
 <div class="col-12 col-md-4 order-md-2">
     <script>
-      var categoryRequest = RestClient.GET('/api/category/list');
+      var categoryRequest = RestClient.GET('/api/category/top');
       categoryRequest.done(function (data) {
         console.log(data);
         var categoriesDom = $('#posting-categories');
